# x/merkledb: serializable hashing type