--- a/x/merkledb/codec_test.go
+++ b/x/merkledb/codec_test.go
@@ -53,12 +53,12 @@ var (
 				children: map[byte]*child{
 					0: {
 						compressedKey: ToKey([]byte{0}),
-						id: ids.ID{
+						id: &serializeable{ids.ID{
 							0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
 							0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
 							0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17,
 							0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f,
-						},
+						}},
 						hasValue: true,
 					},
 				},
@@ -86,22 +86,22 @@ var (
 				children: map[byte]*child{
 					0: {
 						compressedKey: ToKey([]byte{0}),
-						id: ids.ID{
+						id: &serializeable{ids.ID{
 							0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07,
 							0x08, 0x09, 0x0a, 0x0b, 0x0c, 0x0d, 0x0e, 0x0f,
 							0x10, 0x11, 0x12, 0x13, 0x14, 0x15, 0x16, 0x17,
 							0x18, 0x19, 0x1a, 0x1b, 0x1c, 0x1d, 0x1e, 0x1f,
-						},
+						}},
 						hasValue: true,
 					},
 					1: {
 						compressedKey: ToKey([]byte{1, 2, 3}),
-						id: ids.ID{
+						id: &serializeable{ids.ID{
 							0x20, 0x21, 0x22, 0x23, 0x24, 0x25, 0x26, 0x27,
 							0x28, 0x29, 0x2a, 0x2b, 0x2c, 0x2d, 0x2e, 0x2f,
 							0x30, 0x31, 0x32, 0x33, 0x34, 0x35, 0x36, 0x37,
 							0x38, 0x39, 0x3a, 0x3b, 0x3c, 0x3d, 0x3e, 0x3f,
-						},
+						}},
 						hasValue: false,
 					},
 				},
@@ -141,7 +141,7 @@ var (
 				for i := byte(0); i < 16; i++ {
 					n.children[i] = &child{
 						compressedKey: ToKey([]byte{i}),
-						id: ids.ID{
+						id: &serializeable{ids.ID{
 							0x00 + i, 0x01 + i, 0x02 + i, 0x03 + i,
 							0x04 + i, 0x05 + i, 0x06 + i, 0x07 + i,
 							0x08 + i, 0x09 + i, 0x0a + i, 0x0b + i,
@@ -150,7 +150,7 @@ var (
 							0x14 + i, 0x15 + i, 0x16 + i, 0x17 + i,
 							0x18 + i, 0x19 + i, 0x1a + i, 0x1b + i,
 							0x1c + i, 0x1d + i, 0x1e + i, 0x1f + i,
-						},
+						}},
 						hasValue: i%2 == 0,
 					}
 				}
@@ -468,7 +468,7 @@ func FuzzCodecDBNodeCanonical(f *testing.F) {
 		) {
 			require := require.New(t)
 			node := &dbNode{}
-			if err := decodeDBNode(b, node); err != nil {
+			if err := decodeDBNode(DefaultHasher, b, node); err != nil {
 				t.SkipNow()
 			}
 
@@ -508,7 +508,7 @@ func FuzzCodecDBNodeDeterministic(f *testing.F) {
 
 					children[byte(i)] = &child{
 						compressedKey: ToKey(childKeyBytes),
-						id:            childID,
+						id:            &serializeable{childID},
 					}
 				}
 				node := dbNode{
@@ -519,7 +519,7 @@ func FuzzCodecDBNodeDeterministic(f *testing.F) {
 				nodeBytes := encodeDBNode(&node)
 				require.Len(nodeBytes, encodedDBNodeSize(&node))
 				var gotNode dbNode
-				require.NoError(decodeDBNode(nodeBytes, &gotNode))
+				require.NoError(decodeDBNode(DefaultHasher, nodeBytes, &gotNode))
 				require.Equal(node, gotNode)
 
 				nodeBytes2 := encodeDBNode(&gotNode)
@@ -541,7 +541,7 @@ func TestCodecDecodeDBNode_TooShort(t *testing.T) {
 		parsedDBNode  dbNode
 		tooShortBytes = make([]byte, 1)
 	)
-	err := decodeDBNode(tooShortBytes, &parsedDBNode)
+	err := decodeDBNode(DefaultHasher, tooShortBytes, &parsedDBNode)
 	require.ErrorIs(err, io.ErrUnexpectedEOF)
 }
 
@@ -560,7 +560,7 @@ func TestDecodeDBNode(t *testing.T) {
 			require := require.New(t)
 
 			var n dbNode
-			require.NoError(decodeDBNode(test.expectedBytes, &n))
+			require.NoError(decodeDBNode(DefaultHasher, test.expectedBytes, &n))
 			require.Equal(test.n, &n)
 		})
 	}
@@ -627,7 +627,7 @@ func Benchmark_DecodeDBNode(b *testing.B) {
 		b.Run(benchmark.name, func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				var n dbNode
-				err := decodeDBNode(benchmark.expectedBytes, &n)
+				err := decodeDBNode(DefaultHasher, benchmark.expectedBytes, &n)
 				require.NoError(b, err)
 			}
 		})