--- a/x/merkledb/hashing.go
+++ b/x/merkledb/hashing.go
@@ -6,7 +6,6 @@ package merkledb
 import (
 	"crypto/sha256"
 	"encoding/binary"
-	"fmt"
 	"slices"
 
 	"github.com/ava-labs/avalanchego/ids"
@@ -23,23 +22,38 @@ var (
 	DefaultHasher = SHA256Hasher
 )
 
-type SerializableID interface {
-	fmt.Stringer
-	ID() ids.ID
-	Encode(writer *codecWriter)
-	EncodeBytes() []byte
-	Len() int
+const MaxHashLen = 33
+
+type HashDataT [MaxHashLen]byte
+
+type SerializableID struct {
+	data HashDataT
+	idImpl[HashDataT]
 }
 
-type serializeable struct {
-	id ids.ID
+func (id SerializableID) ID() ids.ID                 { return id.idImpl.ID(id.data) }
+func (id SerializableID) Encode(writer *codecWriter) { id.idImpl.Encode(id.data, writer) }
+func (id SerializableID) EncodeBytes() []byte        { return id.idImpl.EncodeBytes(id.data) }
+func (id SerializableID) Len() int                   { return id.idImpl.Len(id.data) }
+func (id SerializableID) String() string             { return id.idImpl.String(id.data) }
+
+type idImpl[T any] interface {
+	String(T) string
+	ID(T) ids.ID
+	Encode(T, *codecWriter)
+	EncodeBytes(T) []byte
+	Len(T) int
 }
 
-func (*serializeable) Len() int                     { return ids.IDLen }
-func (s *serializeable) ID() ids.ID                 { return s.id }
-func (s *serializeable) Encode(writer *codecWriter) { writer.ID(s.id) }
-func (s *serializeable) EncodeBytes() []byte        { return s.id[:] }
-func (s *serializeable) String() string             { return s.id.String() }
+type sIDImpl struct{}
+
+var _ idImpl[HashDataT] = sIDImpl{}
+
+func (sIDImpl) Len(HashDataT) int                          { return ids.IDLen }
+func (sIDImpl) ID(data HashDataT) ids.ID                   { return ids.ID(data[:HashLength]) }
+func (sIDImpl) Encode(data HashDataT, writer *codecWriter) { writer.ID(ids.ID(data[:HashLength])) }
+func (sIDImpl) EncodeBytes(data HashDataT) []byte          { return data[:HashLength] }
+func (s sIDImpl) String(data HashDataT) string             { return s.ID(data).String() }
 
 type Hasher interface {
 	// Returns the canonical hash of the non-nil [node].
@@ -53,26 +67,34 @@ type Hasher interface {
 	Empty() SerializableID
 }
 
+func serializeable(id ids.ID) SerializableID {
+	s := SerializableID{
+		idImpl: sIDImpl{},
+	}
+	copy(s.data[:], id[:])
+	return s
+}
+
 type sha256Hasher struct{}
 
 func (*sha256Hasher) Decode(reader *codecReader) (SerializableID, error) {
 	id, err := reader.ID()
 	if err != nil {
-		return nil, err
+		return SerializableID{}, err
 	}
-	return &serializeable{id}, nil
+	return serializeable(id), nil
 }
 
 func (*sha256Hasher) DecodeBytes(b []byte) (SerializableID, error) {
 	id, err := ids.ToID(b)
 	if err != nil {
-		return nil, err
+		return SerializableID{}, err
 	}
-	return &serializeable{id}, nil
+	return serializeable(id), nil
 }
 
 func (*sha256Hasher) Empty() SerializableID {
-	return &serializeable{ids.Empty}
+	return serializeable(ids.Empty)
 }
 
 // This method is performance critical. It is not expected to perform any memory
@@ -81,10 +103,10 @@ func (*sha256Hasher) HashNode(n *node) SerializableID {
 	var (
 		// sha.Write always returns nil, so we ignore its return values.
 		sha  = sha256.New()
-		hash ids.ID
+		hash = serializeable(ids.Empty)
 		// The hash length is larger than the maximum Uvarint length. This
 		// ensures binary.AppendUvarint doesn't perform any memory allocations.
-		emptyHashBuffer = hash[:0]
+		emptyHashBuffer = hash.data[:0]
 	)
 
 	// By directly calling sha.Write rather than passing sha around as an
@@ -128,7 +150,7 @@ func (*sha256Hasher) HashNode(n *node) SerializableID {
 	_, _ = sha.Write(binary.AppendUvarint(emptyHashBuffer, uint64(n.key.length)))
 	_, _ = sha.Write(n.key.Bytes())
 	sha.Sum(emptyHashBuffer)
-	return &serializeable{hash}
+	return hash
 }
 
 // This method is performance critical. It is not expected to perform any memory