--- a/x/merkledb/proof.go
+++ b/x/merkledb/proof.go
@@ -76,7 +76,7 @@ func (node *ProofNode) ToProto() *pb.ProofNode {
 
 	for childIndex, childID := range node.Children {
 		childID := childID
-		pbNode.Children[uint32(childIndex)] = childID[:]
+		pbNode.Children[uint32(childIndex)] = childID
 	}
 
 	return pbNode