## Why this should be merged
Prep work for eth compatible serialization work (draft [here](https://github.com/ava-labs/avalanchego/pull/3168))

## How this works
Adds an interface "SerializeableID" which provides either a hash (ids.ID) or a serialization (to writer or bytes)
The bytes form is used in proof nodes.

The hasher interface can decode the ID from either a reader or bytes (counterpart to above).
Additionally the hasher must return the empty hash as a serializable ID. (used in addChildWithID? not 100% sure about this).



## How this was tested
CI